requirements.txt
pytelegrambotapi
python-socketio
websocket-client